Refactor clasp push (#514)
* Refactor clone code


* Make urls.ts file and refactor


* Refactor clasp push


* Fix nits


* Fix merge issue
